:recycle: finally super refactoring (#94)
* :recycle: interfaces and classes abstraction awful refactoring

* :rotating_light: stfu u dog

* :rotating_light: i said stfu u dog

* :recycle: splitting commands into several files

* :recycle: stfu u dog again

* :recycle: split messages, take out format command

* :recycle: fix barker warnings

* :recycle: sync messages and handlers packages
