+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/Command.kt
@@ -3,70 +3,98 @@ package com.github.pool_party.pull_party_bot.commands
 import com.elbekD.bot.Bot
 import com.elbekD.bot.types.BotCommand
 import com.elbekD.bot.types.Message
+import com.github.pool_party.pull_party_bot.commands.messages.ON_ADMINS_PARTY_CHANGE
+import com.github.pool_party.pull_party_bot.commands.messages.ON_PERMISSION_DENY
+import com.github.pool_party.pull_party_bot.commands.messages.ON_SENDER_FAIL
+import com.github.pool_party.pull_party_bot.database.dao.ChatDao
 import mu.KotlinLogging
 import java.time.LocalDateTime
 
 private val logger = KotlinLogging.logger {}
 
-fun Bot.registerCommands() {
-    val commands = Command.all
-    commands.forEach {
-        onCommand(it.command) { msg, args ->
-            logger.info {
-                "${LocalDateTime.now()} ${it.command} <- ${msg.from?.username}@\"${msg.chat.title}\": \"${msg.text}\""
-            }
-            it.action.invoke(this, msg, args)
-        }
-    }
-    setMyCommands(commands.map { it.toBotCommand() }).join()
-}
+interface Command {
 
-class Command(
-    val command: String,
-    val description: String,
-    val helpMessage: String,
-    val action: Bot.(Message, String?) -> Unit
-) {
-    init {
-        all.add(this)
-        helpMessages[command.drop(1)] = helpMessage
-    }
+    /**
+     * Command name starting with "/".
+     */
+    val command: String
+
+    /**
+     * Command description that will be displayed in a small popup in Telegram as you type commands.
+     */
+    val description: String
+
+    /**
+     * Command help message that will be displayed on command /help <command>
+     */
+    val helpMessage: String
+
+    fun onMessage(bot: Bot)
 
     fun toBotCommand() = BotCommand(command, description)
-
-    companion object {
-        val all = mutableListOf<Command>()
-        val helpMessages = mutableMapOf<String, String>()
-    }
 }
 
-fun newCommand(commandName: String, description: String, helpMessage: String, action: Bot.(Message, String?) -> Unit) =
-    Command("/$commandName", description, helpMessage, action)
-
-fun newNoArgumentCommand(commandName: String, description: String, helpMessage: String, action: Bot.(Message) -> Unit) =
-    newCommand(commandName, description, helpMessage) { msg, _ -> action(msg) }
-
-fun newAdministratorCommand(
+abstract class AbstractCommand(
     commandName: String,
-    description: String,
-    helpMessage: String,
-    action: Bot.(Message, String?) -> Unit
-) =
-    newCommand(commandName, description, helpMessage) { msg, args ->
-        val sender = msg.from
-        val chatId = msg.chat.id
+    override val description: String,
+    override val helpMessage: String
+) : Command {
+    override val command = "/$commandName"
+
+    abstract fun Bot.action(message: Message, args: String?)
+
+    override fun onMessage(bot: Bot) = bot.onCommand(command) { message, args ->
+        logger.info {
+            "${LocalDateTime.now()} $command <- ${message.from?.username}@${message.chat.title}: \"${message.text}\""
+        }
+        bot.action(message, args)
+    }
+
+    protected fun Bot.modifyCommandAssertion(chatId: Long, name: String): Boolean =
+        (name == "admins").not().also { if (!it) sendMessage(chatId, ON_ADMINS_PARTY_CHANGE, "Markdown") }
+
+    protected fun parseArgs(args: String?): List<String>? =
+        args?.split(' ')?.map { it.trim().toLowerCase() }?.filter { it.isNotBlank() }
+}
+
+abstract class CaseCommand(command: String, description: String, helpMessage: String, protected val chatDao: ChatDao) :
+    AbstractCommand(command, description, helpMessage) {
+
+    protected fun Bot.sendCaseMessage(
+        chatId: Long,
+        message: String,
+        parseMode: String? = null,
+        replyTo: Int? = null
+    ) =
+        sendMessage(
+            chatId,
+            if (chatDao.getRude(chatId)) message.toUpperCase() else message,
+            parseMode,
+            replyTo = replyTo
+        )
+}
+
+abstract class AdministratorCommand(command: String, description: String, helpMessage: String, chatDao: ChatDao) :
+    CaseCommand(command, description, helpMessage, chatDao) {
+
+    abstract fun Bot.mainAction(message: Message, args: String?)
+
+    override fun Bot.action(message: Message, args: String?) {
+        val sender = message.from
+        val chatId = message.chat.id
         if (sender == null) {
             sendMessage(chatId, ON_SENDER_FAIL, "Markdown")
-            return@newCommand
+            return
         }
 
-        val chatType = msg.chat.type
+        val chatType = message.chat.type
         if ((chatType == "group" || chatType == "supergroup") &&
             getChatAdministrators(chatId).join().all { it.user != sender }
         ) {
             sendMessage(chatId, ON_PERMISSION_DENY, "Markdown")
-            return@newCommand
+            return
         }
 
-        action(msg, args)
+        mainAction(message, args)
     }
+}
